+++ droidboy/src/com/appboy/sample/CustomLoggingDialog.java
@@ -11,13 +11,19 @@ import android.widget.Toast;
 
 import com.appboy.Appboy;
 import com.appboy.AppboyUser;
+import com.appboy.models.outgoing.AppboyProperties;
 import com.appboy.ui.support.StringUtils;
 import java.math.BigDecimal;
+import java.util.Date;
 
 public class CustomLoggingDialog extends DialogPreference {
   private EditText mCustomAttributeKey;
   private EditText mCustomAttributeValue;
   private EditText mCustomEventName;
+  private EditText mCustomEventPropertyKey;
+  private EditText mCustomEventPropertyValue;
+  private EditText mCustomPurchasePropertyKey;
+  private EditText mCustomPurchasePropertyValue;
   private EditText mCustomPurchaseName;
   private EditText mCustomPurchaseQuantity;
   private EditText mCustomAttributeArrayKey;
@@ -36,8 +42,12 @@ public class CustomLoggingDialog extends DialogPreference {
     mCustomAttributeKey = (EditText) view.findViewById(R.id.custom_attribute_key);
     mCustomAttributeValue = (EditText) view.findViewById(R.id.custom_attribute_value);
     mCustomEventName = (EditText) view.findViewById(R.id.custom_event);
+    mCustomEventPropertyKey =  (EditText) view.findViewById(R.id.custom_event_property_key);
+    mCustomEventPropertyValue =  (EditText) view.findViewById(R.id.custom_event_property_value);
     mCustomPurchaseName = (EditText) view.findViewById(R.id.custom_purchase);
     mCustomPurchaseQuantity = (EditText) view.findViewById(R.id.purchase_qty);
+    mCustomPurchasePropertyKey =  (EditText) view.findViewById(R.id.purchase_property_key);
+    mCustomPurchasePropertyValue =  (EditText) view.findViewById(R.id.purchase_property_value);
     mCustomAttributeArrayKey = (EditText) view.findViewById(R.id.custom_attribute_array_key);
     mCustomAttributeArrayValue = (EditText) view.findViewById(R.id.custom_attribute_array_value);
     mCustomAttributeArrayChoices = (RadioGroup) view.findViewById(R.id.custom_attribute_array_radio);
@@ -56,8 +66,12 @@ public class CustomLoggingDialog extends DialogPreference {
       String customAttributeKeyName = mCustomAttributeKey.getText().toString();
       String customAttributeValueName = mCustomAttributeValue.getText().toString();
       String customEventName = mCustomEventName.getText().toString();
+      String customEventPropertyKey = mCustomEventPropertyKey.getText().toString();
+      String customEventPropertyValue = mCustomEventPropertyValue.getText().toString();
       String customPurchaseName = mCustomPurchaseName.getText().toString();
       String customPurchaseQuantity = mCustomPurchaseQuantity.getText().toString();
+      String customPurchasePropertyKey = mCustomPurchasePropertyKey.getText().toString();
+      String customPurchasePropertyValue = mCustomPurchasePropertyValue.getText().toString();
       String customAttributeArrayKey = mCustomAttributeArrayKey.getText().toString();
       String customAttributeArrayValue = mCustomAttributeArrayValue.getText().toString();
       int attributeArrayResourceId = mCustomAttributeArrayChoices.getCheckedRadioButtonId();
@@ -73,14 +87,47 @@ public class CustomLoggingDialog extends DialogPreference {
         notifyResult(appboyUser.setCustomUserAttribute(customAttributeKeyName, customAttributeValueName), "user attribute! key=" + customAttributeKeyName + ", value=" + customAttributeValueName);
       }
       if (!StringUtils.isNullOrBlank(customEventName)) {
-        notifyResult(Appboy.getInstance(getContext()).logCustomEvent(customEventName), "custom event: " + customEventName);
+        if (!StringUtils.isNullOrBlank(customEventPropertyKey)) {
+          AppboyProperties eventProperties = new AppboyProperties();
+          if (StringUtils.isNullOrBlank(customEventPropertyValue)) {
+            customEventPropertyValue = "default";
+          }
+          eventProperties.addProperty(customEventPropertyKey, customEventPropertyValue);
+          eventProperties.addProperty("time", new Date(System.currentTimeMillis()));
+          eventProperties.addProperty("boolean", false);
+          eventProperties.addProperty("double", 2.5);
+          eventProperties.addProperty("integer", 3);
+          eventProperties.addProperty("string", "string");
+          notifyResult(Appboy.getInstance(getContext()).logCustomEvent(customEventName, eventProperties),
+              String.format("custom event: %s. Custom properties: %s, %s",customEventName, customEventPropertyKey, customEventPropertyValue));
+        } else {
+          notifyResult(Appboy.getInstance(getContext()).logCustomEvent(customEventName), "custom event: " + customEventName);
+        }
       }
       if (!StringUtils.isNullOrBlank(customPurchaseName)) {
-        if (!StringUtils.isNullOrBlank(customPurchaseQuantity)) {
+        if (StringUtils.isNullOrBlank(customPurchaseQuantity) && StringUtils.isNullOrBlank(customPurchasePropertyKey)) {
+          notifyResult(Appboy.getInstance(getContext()).logPurchase(customPurchaseName, "USD", BigDecimal.ONE), "single purchase of: " + customPurchaseName);
+        } else if (StringUtils.isNullOrBlank(customPurchasePropertyKey)) {
           notifyResult(Appboy.getInstance(getContext()).logPurchase(customPurchaseName, "USD", BigDecimal.ONE, Integer.parseInt(customPurchaseQuantity)),
               customPurchaseQuantity + " purchases of: " + customPurchaseName);
         } else {
-          notifyResult(Appboy.getInstance(getContext()).logPurchase(customPurchaseName, "USD", BigDecimal.ONE), "single purchase of: " + customPurchaseName);
+          AppboyProperties purchaseProperties = new AppboyProperties();
+          if (StringUtils.isNullOrBlank(customPurchasePropertyValue)) {
+            customPurchasePropertyValue = "default";
+          }
+          purchaseProperties.addProperty(customPurchasePropertyKey, customPurchasePropertyValue);
+          purchaseProperties.addProperty("time", new Date(System.currentTimeMillis()));
+          purchaseProperties.addProperty("boolean", true);
+          purchaseProperties.addProperty("double", 1.5);
+          purchaseProperties.addProperty("integer", 2);
+          purchaseProperties.addProperty("string", "string");
+          if (StringUtils.isNullOrBlank(customPurchaseQuantity)) {
+            notifyResult(Appboy.getInstance(getContext()).logPurchase(customPurchaseName, "USD", BigDecimal.ONE, purchaseProperties),
+                String.format("single purchase: %s. Custom properties: %s, %s", customPurchaseName, customPurchasePropertyKey, customPurchasePropertyValue));
+          } else {
+            notifyResult(Appboy.getInstance(getContext()).logPurchase(customPurchaseName, "USD", BigDecimal.ONE, Integer.parseInt(customPurchaseQuantity), purchaseProperties),
+                String.format("%s purchases of %s. Custom properties: %s, %s", customPurchaseQuantity, customPurchaseName, customPurchasePropertyKey, customPurchasePropertyValue));
+          }
         }
       }
       if (!StringUtils.isNullOrBlank(customAttributeArrayKey)) {